improved customer management with payment infos and unlimited addresses

+++ application/modules/contacts/controllers/AddressController.php
@@ -2,6 +2,12 @@
 
 class Contacts_AddressController extends Zend_Controller_Action
 {
+	public function init()
+	{
+		$this->view->client = Zend_Registry::get('Client');
+		$this->view->user = $this->_user = Zend_Registry::get('User');
+	}
+
 	public function addAction()
 	{
 		$request = $this->getRequest();
@@ -10,15 +16,37 @@ class Contacts_AddressController extends Zend_Controller_Action
 		$this->_helper->getHelper('layout')->disableLayout();
 
 		$form = new Contacts_Form_Contact();
+		$options = $this->_helper->Options->getOptions($form, $this->_user['clientid']);
+	    $this->view->options = $options;
+	    $this->view->action = 'add';
+
+		$client = Zend_Registry::get('Client');
 
 		if($request->isPost()) {
 			$data = $request->getPost();
-			if($form->isValid($data) || true) {
-				$addressDb = new Contacts_Model_DbTable_Address();
-				$latest = end($addressDb->getAddress($data['contactid']));
-				$addressDb->addAddress($data['contactid'], $data['type'], '', $latest['ordering']+1);
-				$address = end($addressDb->getAddress($data['contactid']));
-				echo $this->view->MultiForm('address', $address);
+			if($data['contactid']) {
+				if($form->isValid($data) || true) {
+					$addressDb = new Contacts_Model_DbTable_Address();
+					$addressDataBefore = $addressDb->getAddress($data['contactid']);
+					$latest = end($addressDataBefore);
+					$addressDb->addAddress($data['contactid'], $data['type'], '', '', '', $client['country'], $latest['ordering']+1);
+					$addressDataAfter = $addressDb->getAddress($data['contactid']);
+					$address = end($addressDataAfter);
+					echo $this->view->MultiForm('address', $address, array(
+                                                                        //array('label' => 'CONTACTS_NAME', 'field' => 'name1'),
+                                                                        array('label' => 'CONTACTS_STREET', 'field' => 'street'),
+                                                                        array('label' => 'CONTACTS_POSTCODE_CITY', 'fields' => array('postcode', 'city')),
+                                                                        array('label' => 'CONTACTS_COUNTRY_ADDRESS_TYPE', 'fields' => array('country', 'type'))
+                                                                        ));
+			    }
+			} else {
+				$timestamp = time();
+				$address = array('id' => $timestamp, 'ordering' => $timestamp, 'type' => 'address', 'address' => '');
+				echo $this->view->MultiForm('address', $address, array(
+                                                                    array('label' => 'CONTACTS_STREET', 'field' => 'street'),
+                                                                    array('label' => 'CONTACTS_POSTCODE_CITY', 'fields' => array('postcode', 'city')),
+                                                                    array('label' => 'CONTACTS_COUNTRY_ADDRESS_TYPE', 'fields' => array('country', 'type'))
+                                                                    ));
 			}
 		}
 	}
@@ -51,11 +79,11 @@ class Contacts_AddressController extends Zend_Controller_Action
 		$this->_helper->viewRenderer->setNoRender();
 		$this->_helper->getHelper('layout')->disableLayout();
 
-		/*if($this->getRequest()->isPost()) {
+		if($this->getRequest()->isPost()) {
 			$id = $this->_getParam('id', 0);
 			$addressDb = new Contacts_Model_DbTable_Address();
 			$addressDb->deleteAddress($id);
-		}*/
+		}
 		//$this->_flashMessenger->addMessage('MESSAGES_SUCCESFULLY_DELETED');
 	}
 }
